Moving Permission Checks to systemCommands
